Update UnusualHeaders.py
Updated the req/res filter logic a little...

We don't want to assess any injected headers from Burp or other plugins as these are likely to include false positives. But we do want to assess request headers that are generated by the browser and proxied via Burp. So for example, if an Active Scan is the trigger of the request, just simply return without doing anything. All responses should be scanned regardless of their trigger. So therefore, skip any request that was not initiated through the Proxy tab...